Use php:apache instead of whippet-server:tomdxw
