+++ docker-images/whippet-wordpress/mu-plugins/mailcatcher.php
