+++ docker-images/whippet-wordpress/php.ini
@@ -1,0 +1,3 @@
+error_reporting=E_ALL
+display_errors=Off
+log_errors=On
