+++ docker-images/whippet-wordpress/wordpress.conf
@@ -1,0 +1,15 @@
+<Directory />
+        RewriteEngine On
+        RewriteBase /
+        RewriteRule ^index\.php$ - [L]
+
+        # add a trailing slash to /wp-admin
+        RewriteRule ^([_0-9a-zA-Z-]+/)?wp-admin$ $1wp-admin/ [R=301,L]
+
+        RewriteCond %{REQUEST_FILENAME} -f [OR]
+        RewriteCond %{REQUEST_FILENAME} -d
+        RewriteRule ^ - [L]
+        RewriteRule ^([_0-9a-zA-Z-]+/)?(wp-(content|admin|includes).*) $2 [L]
+        RewriteRule ^([_0-9a-zA-Z-]+/)?(.*\.php)$ $2 [L]
+        RewriteRule . index.php [L]
+</Directory>
